@@ -1,6 +1,6 @@
 {
   "name": "3nweb",
-  "version": "0.19.3",
+  "version": "0.19.8",
   "private": true,
   "description": "3NWeb platform client for desktop environments",
   "main": "build/all/main.js",
@@ -39,7 +39,7 @@
     "adm-zip": "^0.5.5",
     "command-line-args": "^5.2.1",
     "command-line-usage": "^6.1.3",
-    "core-3nweb-client-lib": "^0.30.3",
+    "core-3nweb-client-lib": "^0.30.4",
     "electron-localshortcut": "^3.2.1",
     "electron-updater": "^4.3.9",
     "mime": "^1.3.4",
@@ -56,6 +56,6 @@
     "electron-build-env": "^0.2.0",
     "electron-builder": "^24.6.4",
     "js-yaml": "^4.1.0",
-    "typescript": "^4.7.4"
+    "typescript": "^5.5.3"
   }
 }
