@@ -9,7 +9,7 @@
   },
   "devDependencies": {
     "@types/jasmine": "^3.9.1",
-    "typescript": "~4.1.5"
+    "typescript": "^5.5.3"
   },
   "author": {
     "name": "3NSoft Inc.",
